requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
asyncpgsa==0.27.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.112.2
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.4
yarl==1.9.4
sqladmin
